this is the htaccess file READ INSTRUCTIONS
this file needs to be at the root of the site. If there is a WordPress htaccess file then the code from this one needs to be included in that file outside of their code- DO NOT TOUCH THE WORDPRESS CODE. If their is no htaccess file this can be used but renamed simply  .htaccess and put at the root of the site (not in any folders). This is a hidden file and it tells the browser to go from the old site to this new one and it keeps all the SEO that has accumulated on the old site to the new one.